requirements.txt
aiohttp>=3.9.0  # For async HTTP requests to OpenAI API
python-dotenv>=1.0.0  # For loading environment variables
typing-extensions>=4.8.0  # For additional type hints 